src/test/java/lotto/domain/model/LottoTest.java
package lotto.domain.model;

import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;

import java.util.Arrays;
import java.util.List;

import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
import static org.assertj.core.api.Assertions.assertThatThrownBy;
import static org.junit.jupiter.api.Assertions.*;

class LottoTest {
	@DisplayName("로또 번호의 개수가 6개가 넘어가면 예외가 발생한다.")
	@Test
	void createLottoByOverSize() {
		assertThatThrownBy(() -> new Lotto(List.of(1, 2, 3, 4, 5, 6, 7)))
				.isInstanceOf(IllegalArgumentException.class);
	}

	@DisplayName("로또 번호에 중복된 숫자가 있으면 예외가 발생한다.")
	@Test
	void createLottoByDuplicatedNumber() {
		// TODO: 이 테스트가 통과할 수 있게 구현 코드 작성
		assertThatThrownBy(() -> new Lotto(List.of(1, 2, 3, 4, 5, 5)))
				.isInstanceOf(IllegalArgumentException.class);
	}

	@Test
	void 제대로_동작() {
		Lotto lotto = new Lotto(Arrays.asList(1, 2, 3, 4, 5, 6));
		List<Integer> answer = Arrays.asList(1, 2, 3, 4, 5, 6);

		assertEquals(answer, lotto.getLotto());
	}

	@Test
	void 예외처리_크기() {
		assertSimpleTest(() ->
				assertThatThrownBy(() -> new Lotto(Arrays.asList(1, 2, 3, 4, 5, 6, 7)))
						.isInstanceOf(IllegalArgumentException.class)
		);
	}

	@Test
	void 예외처리_크기2() {
		assertSimpleTest(() ->
				assertThatThrownBy(() -> new Lotto(Arrays.asList(1, 2, 3))).isInstanceOf(IllegalArgumentException.class)
		);
	}
}